Module rustdsp::components

Modules

interpolator

loop_filter

The loop filter is a critical tool to implement advance advance systems and digital circuits a good resource for this implementation is here https://ieeexplore.ieee.org/document/6154200

lpf

low pass filters (LPFs) are a critical component in filtering and used in the construction of larger, more powerful components. Implementation based off https://www.dsp-weimich.com/digital-signal-processing/iir-first-order-digital-filter/

match_filter

nco

Voltage Controlled Oscillator (NCO) is a major component in the construction of loops. Since voltage doesn’t exist in software, we create a impression of this with a fake voltage to adjust the oscillation. A positive voltage increases the

pi_loop

resampler

timing_error_detection

Functions

threshold

this will take a sample and either set it to -1.0 or 1.0 .