Trait rustdsp::elements::element::Element

pub trait Element: ElementClone + Send {
    // Required methods
    fn init(
        &mut self,
        builder: &mut WorkflowBuilder,
        samples: &mut ElementParameter
    );
    fn run(&mut self, samples: &mut ElementParameter);
    fn halt(&self) -> bool;
    fn stop(&self, samples: &mut ElementParameter) -> bool;
    fn is_source(&self) -> bool;
}

Required Methods

fn init( &mut self, builder: &mut WorkflowBuilder, samples: &mut ElementParameter )

This is the main workflow setup. This runs before build window. Samples is the input and output buffer

fn run(&mut self, samples: &mut ElementParameter)

This will run if halt is returns “true”

fn halt(&self) -> bool

This tells the pipeline builder to stop and run the “run” function. You should only do this if you absolutely need to

fn stop(&self, samples: &mut ElementParameter) -> bool

This tells the pipeline during run time that this is the last element in the pipeline and to stop here

fn is_source(&self) -> bool

This will tell the workflow builder that this block is the top part in a workflow chain (like signal generators) and doesn’t take an input