Function rustdsp::elements::prefabs::nco::add_nco

pub fn add_nco(
    builder: &mut WorkflowBuilder,
    sps: usize,
    sample_rate: f32,
    frequency: f32,
    frequency_error_scalar: &ElementParameter
) -> ComplexF32

This will add a NCO component to the builder. This takes in the samples per a symbol (sps) and the error and adjusts the components. This will return the array where the complex values are sent